rename http directory
