+++ packages/serverpod_client/lib/src/serverpod_client_shared_private.dart
@@ -2,7 +2,8 @@ import 'package:serverpod_serialization/serverpod_serialization.dart';
 
 import 'serverpod_client_exception.dart';
 
-import './_sdk/http_status_html.dart' if (dart.library.io) './_sdk/http_status_io.dart';
+import 'http/http_status_html.dart'
+    if (dart.library.io) 'http/http_status_io.dart';
 
 /// Encodes arguments for serialization.
 String formatArgs(
